bots/requirements.txt
weaviate-client>=4.0.0
requests>=2.31.0
packaging>=24.0
